Create README.md
Objective :
To develop an Machine Learning Model to predict Hill and Valley using Logistic Regression method that can accurately classify geographical locations as either hills or valleys based on a set of input features. The model will be trained on a labeled dataset of geographical features and their corresponding classifications, and then evaluated on a separate test dataset to measure its performance. The ultimate goal of this project is to provide a useful tool for identifying hills and valleys in various geographic locations, which can have important applications in fields such as geology, agriculture, and urban planning.